refactor(tests): replace import of strip_ansi_codes in test files
Replaced import of strip_ansi_codes from tests.cli_fixtures to tests.utils in multiple test files for consistency and better organization.
